script: Move individual interfaces into subfolders
All of these are used in 1 (or very few) place. Therefore,
move them to the subfolder where they are used.

Part of #38901

Testing: it compiles.
