+++ components/script/dom/document/mod.rs
@@ -13,5 +13,6 @@ pub(crate) mod documenttype;
 pub(crate) mod focus;
 mod tree_ordered_index_map;
 pub(crate) mod visibilitystateentry;
+pub(crate) mod websocket;
 
 pub(crate) use self::document::*;
