+++ components/script/dom/globalscope/mod.rs
@@ -2,6 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+pub(crate) mod broadcastchannel;
 pub(crate) use self::globalscope::*;
 #[expect(clippy::module_inception, reason = "The interface name is GlobalScope")]
 pub(crate) mod globalscope;
