+++ components/script/dom/mod.rs
@@ -222,7 +222,6 @@ pub(crate) mod bindings;
 pub(crate) mod bluetooth;
 #[cfg(feature = "bluetooth")]
 pub(crate) use self::bluetooth::*;
-pub(crate) mod broadcastchannel;
 mod canvas;
 pub(crate) use self::canvas::*;
 pub(crate) mod characterdata;
@@ -293,12 +292,10 @@ pub(crate) mod navigator;
 pub(crate) use self::navigator::*;
 pub(crate) mod node;
 pub(crate) use self::node::*;
-pub(crate) mod notification;
 pub(crate) mod performance;
 pub(crate) use self::performance::*;
 pub(crate) mod permission;
 pub(crate) use self::permission::*;
-pub(crate) mod processingoptions;
 pub(crate) mod promise;
 pub(crate) use self::promise::*;
 pub(crate) mod quotaexceedederror;
@@ -309,8 +306,6 @@ pub(crate) mod reporting;
 pub(crate) use self::reporting::*;
 pub(crate) mod resizeobserver;
 pub(crate) use self::resizeobserver::*;
-pub(crate) mod screen;
-mod scrolling_box;
 pub(crate) mod security;
 pub(crate) use self::security::*;
 pub(crate) mod selection;
@@ -337,7 +332,6 @@ pub(crate) mod trustedtypes;
 pub(crate) use self::trustedtypes::*;
 pub(crate) mod url;
 pub(crate) use self::url::*;
-pub(crate) mod useractivation;
 pub(crate) mod userscripts;
 pub(crate) mod values;
 pub(crate) mod visualviewport;
@@ -346,7 +340,6 @@ pub(crate) use self::wakelock::*;
 pub(crate) mod webgl;
 pub(crate) use self::webgl::extensions::ext::*;
 pub(crate) use self::webgl::*;
-pub(crate) mod websocket;
 #[cfg(feature = "webxr")]
 mod webxr;
 #[cfg(feature = "webxr")]
